refractor nonogramData

+++ types/lineId.go
@@ -2,5 +2,5 @@ package types
 
 type LineID struct {
 	Direction Direction
-	Index     uint8
+	Index     int
 }
